Title: 낮은 지점들, Time: 1243ms, Memory: 44MB, Status: Passed - CodeTree

+++ 250305/낮은 지점들/lowest-points.java
@@ -5,20 +5,20 @@ import java.util.Scanner;
 public class Main {
 
 	public static int n;
-	public static Map<Integer, Integer> map = new HashMap<>();
+	public static Map<Long, Long> map = new HashMap<>();
 	public static Scanner sc = new Scanner(System.in);
 
 	public static void main(String[] args) {
 		n = sc.nextInt();
 
 		for (int i = 0; i < n; i++) {
-			int x = sc.nextInt();
-			int y = sc.nextInt();
+			long x = sc.nextLong();
+			long y = sc.nextLong();
 
-			map.put(x, Math.min(map.getOrDefault(x, Integer.MAX_VALUE), y));
+			map.put(x, Math.min(map.getOrDefault(x, Long.MAX_VALUE), y));
 		}
 
-		Integer sum = map.values().stream().reduce((a, b) -> a + b).get();
+		Long sum = map.values().stream().reduce((a, b) -> a + b).get();
 		System.out.println(sum);
 	}
 }
